pyproject.toml
[project]
name = "dreaminamonitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.123.8",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
